+++ tests/test_data/style_mismatch/this_can_be_any_style.py
@@ -1,0 +1,8 @@
+def myFunc(arg1: str):
+    """
+    This is a docstring.
+
+    This function does something very fancy. Yes, it does! But I don't think
+    :class:`something` so I can remember what it does.
+    """
+    pass